requirements.txt
streamlit>=1.30.0
langchain==0.1.16
langchain-community==0.0.32
langchain-core==0.1.45
langchain-text-splitters==0.0.1
langchain-groq==0.1.3
faiss-cpu
sentence-transformers>=2.2.2
transformers>=4.36.0
huggingface-hub>=0.20.0
sendgrid
python-dotenv
supabase
pandas
pymupdf
